+++ Kooboo.Sites/Scripting/Global/Database/DynamicTableObject.cs
@@ -4,28 +4,21 @@ using Kooboo.IndexedDB.Dynamic;
 using Kooboo.Sites.Extensions;
 using Kooboo.Sites.Models;
 using Kooboo.Sites.Repository;
+using Kooboo.Sites.Scripting.Global.Database;
 using Kooboo.Sites.Scripting.Helper;
 using System.Collections.Generic;
-using System.Linq; 
+using System.Linq;
 
 namespace KScript
 {
     [Newtonsoft.Json.JsonConverter(typeof(JsonConverterDynamicObject))]
-    public class DynamicTableObject :IDynamicTableObject
+    public class DynamicTableObject : DynamicTableObjectBase, IDynamicTableObject
     {
         [KIgnore]
         public IDictionary<string, object> obj { get; set; }
 
         private RenderContext context { get; set; }
         private Table table { get; set; }
-
-        public Dictionary<string, object> Values
-        {
-            get
-            {
-                return this.obj.ToDictionary(o => o.Key, o => o.Value);
-            }
-        }
 
         public DynamicTableObject(IDictionary<string, object> orgObj, Table orgtable, RenderContext renderContext)
         {
@@ -34,20 +27,7 @@ namespace KScript
             this.table = orgtable;
         }
 
-        [KIgnore]
-        public object this[string key]
-        {
-            get
-            {
-                return GetValueFromDict(key);
-            }
-            set
-            {
-                this.obj[key] = value;
-            }
-        }
-
-        private object GetValueFromDict(string key)
+        internal override object GetValueFromDict(string key)
         {
             if (obj.ContainsKey(key))
             {
@@ -133,21 +113,6 @@ namespace KScript
             }
             return null;
 
-        }
-
-        public object GetValue(string FieldName)
-        {
-            return GetValueFromDict(FieldName);
-        }
-
-        public object GetValue(string FieldName, RenderContext Context)
-        {
-            return GetValueFromDict(FieldName);
-        }
-
-        public void SetValue(string FieldName, object Value)
-        {
-            obj[FieldName] = Value;
         }
     }
 }
